+++ UserRegistrationProfileManagement/src/main/java/com/cts/lms/service/UserServiceImpl.java
@@ -1,11 +1,8 @@
 package com.cts.lms.service;
-
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.cts.lms.exceptions.UserNotFound;
 import com.cts.lms.model.User;
 import com.cts.lms.repository.UserRepository;
@@ -41,10 +38,35 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public String deleteUser(int userId) {
-		repository.deleteById(userId);
-		return "User Details deleted Successfully!!!";
+	public String deleteUser(int userId) throws UserNotFound{
+		Optional<User> optional = repository.findById(userId);
+		if (optional.isPresent()) {
+			repository.deleteById(userId);
+			return "User Details deleted Successfully!!!";
+		}
+		else {
+			throw new UserNotFound("Invalid Course ID!!!");
+		}
+	}
+	
+	@Override
+	public User getUserById(int userId) throws UserNotFound {
+		
+		Optional<User> optional = repository.findById(userId);
+		if (optional.isPresent())
+			return optional.get();
+		else
+			throw new UserNotFound("User Id is Invalid...");
+
 	}
 
+	@Override
+	public Boolean checkUserExist(int userId) throws UserNotFound {
+		boolean response = repository.existsById(userId);
+		if (response)
+			return response;
+		else
+			throw new UserNotFound("User Id is Invalid...");
+	}
 	
 }
